+++ serenity-stats/src/main/kotlin/net/serenitybdd/reports/model/TestResultSummary.kt
@@ -1,13 +1,14 @@
 package net.serenitybdd.reports.model
 
-class TestResultSummary(val totalCount: Int,
-                        val countByResult: Map<String, Long>,
-                        val percentageByResult: Map<String, Int>,
-                        val totalTestDuration: String,
-                        val clockTestDuration: String,
-                        val averageTestDuration: String,
-                        val maxTestDuration: String,
-                        val minTestDuration: String
+class TestResultSummary(
+    val totalCount: Long,
+    val countByResult: Map<String, Long>,
+    val percentageByResult: Map<String, Int>,
+    val totalTestDuration: String,
+    val clockTestDuration: String,
+    val averageTestDuration: String,
+    val maxTestDuration: String,
+    val minTestDuration: String
                         )
